requirements.txt
Flask==2.2.3
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0
